basic resynthesis of 303-ish waves
the problem now is that i've parameterized the real part of the s-domain
poles and zeros as an exponential for numerical stability when
optimizing... but when resynthesizing, i need to not do that so that it
decays correctly with the imaginary part.

basically the poles (and zero) have a base position (in the Laplace
domain) which is then multiplied by the cutoff frequency, which is
exponentially decaying. those are then transformed to the z domain.